Refactor drills UI helpers into partial classes
