+++ TFTV/TFTVDrills/CanvasGroupExtensions.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace TFTV.TFTVDrills
+{
+    internal static class CanvasGroupExtensions
+    {
+        public static void SetAlpha(this CanvasGroup group, float alpha)
+        {
+            if (group != null)
+            {
+                group.alpha = alpha;
+            }
+        }
+    }
+}
